chore: updated item quantity from integers to double

+++ src/test/java/isep/mock/DistributionNetworkWithOrdersMock.java
@@ -26,28 +26,28 @@ public class DistributionNetworkWithOrdersMock {
     DailyData dataProducer1 = new DailyData();
     DailyData dataProducer2 = new DailyData();
 
-    Map<Product, Integer> mapProducer = new HashMap<>();
-    mapProducer.put(new Product("banana"), 100);
-    mapProducer.put(new Product("orange"), 200);
-    mapProducer.put(new Product("limon"), 300);
+    Map<Product, Double> mapProducer = new HashMap<>();
+    mapProducer.put(new Product("banana"), 100.);
+    mapProducer.put(new Product("orange"), 200.);
+    mapProducer.put(new Product("limon"), 300.);
 
     dataProducer.addDayData(1, mapProducer);
     dataProducer.addDayData(2, mapProducer);
     dataProducer.addDayData(3, mapProducer);
 
-    Map<Product, Integer> mapProducer1 = new HashMap<>();
-    mapProducer.put(new Product("banana"), 300);
-    mapProducer.put(new Product("orange"), 200);
-    mapProducer.put(new Product("limon"), 100);
+    Map<Product, Double> mapProducer1 = new HashMap<>();
+    mapProducer.put(new Product("banana"), 300.);
+    mapProducer.put(new Product("orange"), 200.);
+    mapProducer.put(new Product("limon"), 100.);
 
     dataProducer1.addDayData(1, mapProducer1);
     dataProducer1.addDayData(2, mapProducer1);
     dataProducer1.addDayData(3, mapProducer1);
 
-    Map<Product, Integer> mapProducer2 = new HashMap<>();
-    mapProducer.put(new Product("banana"), 300);
-    mapProducer.put(new Product("orange"), 200);
-    mapProducer.put(new Product("limon"), 100);
+    Map<Product, Double> mapProducer2 = new HashMap<>();
+    mapProducer.put(new Product("banana"), 300.);
+    mapProducer.put(new Product("orange"), 200.);
+    mapProducer.put(new Product("limon"), 100.);
 
     dataProducer2.addDayData(1, mapProducer2);
     dataProducer2.addDayData(2, mapProducer2);
@@ -61,21 +61,21 @@ public class DistributionNetworkWithOrdersMock {
     DailyData dataClient1 = new DailyData();
     DailyData dataClient2 = new DailyData();
 
-    Map<Product, Integer> mapClient = new HashMap<>();
-    mapClient.put(new Product("banana"), 100);
+    Map<Product, Double> mapClient = new HashMap<>();
+    mapClient.put(new Product("banana"), 100.);
 
     dataClient.addDayData(1, mapClient);
 
-    Map<Product, Integer> mapClient1 = new HashMap<>();
-    mapClient1.put(new Product("orange"), 200);
-    mapClient1.put(new Product("limon"), 100);
+    Map<Product, Double> mapClient1 = new HashMap<>();
+    mapClient1.put(new Product("orange"), 200.);
+    mapClient1.put(new Product("limon"), 100.);
 
     dataClient1.addDayData(1, mapClient1);
 
-    Map<Product, Integer> mapClient2 = new HashMap<>();
-    mapClient2.put(new Product("orange"), 200);
-    mapClient2.put(new Product("limon"), 100);
-    mapClient2.put(new Product("banana"), 150);
+    Map<Product, Double> mapClient2 = new HashMap<>();
+    mapClient2.put(new Product("orange"), 200.);
+    mapClient2.put(new Product("limon"), 100.);
+    mapClient2.put(new Product("banana"), 150.);
 
     dataClient2.addDayData(3, mapClient2);
